Clean up existing tests
