+++ http/netty/components/org.wso2.carbon.transport.http.netty/src/test/java/org/wso2/carbon/transport/http/netty/contentaware/EchoMessageProcessor.java
@@ -36,8 +36,8 @@ import java.util.concurrent.Executors;
 /**
  * A Message processor which echos the incoming message
  */
-public class MessageEchoingMessageProcessor implements CarbonMessageProcessor {
-    private static final Logger logger = LoggerFactory.getLogger(MessageEchoingMessageProcessor.class);
+public class EchoMessageProcessor implements CarbonMessageProcessor {
+    private static final Logger logger = LoggerFactory.getLogger(EchoMessageProcessor.class);
 
     private TransportSender transportSender;
