+++ http/netty/components/org.wso2.carbon.transport.http.netty/src/test/java/org/wso2/carbon/transport/http/netty/passthrough/PassThroughHttpTestCase.java
@@ -21,8 +21,12 @@ package org.wso2.carbon.transport.http.netty.passthrough;
 import io.netty.handler.codec.http.HttpMethod;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.wso2.carbon.transport.http.netty.config.ListenerConfiguration;
-import org.wso2.carbon.transport.http.netty.config.SenderConfiguration;
+import org.testng.annotations.AfterClass;
+import org.testng.annotations.BeforeClass;
+import org.testng.annotations.Test;
+import org.wso2.carbon.transport.http.netty.common.Constants;
+import org.wso2.carbon.transport.http.netty.config.TransportsConfiguration;
+import org.wso2.carbon.transport.http.netty.config.YAMLTransportConfigurationBuilder;
 import org.wso2.carbon.transport.http.netty.listener.HTTPTransportListener;
 import org.wso2.carbon.transport.http.netty.util.TestUtil;
 import org.wso2.carbon.transport.http.netty.util.server.HTTPServer;
@@ -32,58 +36,55 @@ import java.net.HttpURLConnection;
 import java.net.URI;
 
 import static org.testng.AssertJUnit.assertEquals;
-import static org.testng.AssertJUnit.assertTrue;
 
 /**
  * A test class for passthrough transport
  */
-public class PassThroughHttpPOSTMethodTestCase {
+public class PassThroughHttpTestCase {
 
-    protected static final Logger LOGGER = LoggerFactory.getLogger(PassThroughHttpPOSTMethodTestCase.class);
+    private static final Logger log = LoggerFactory.getLogger(PassThroughHttpTestCase.class);
 
     private HTTPTransportListener httpTransportListener;
 
+    private static final String testValue = "Test Message";
+
     private HTTPServer httpServer;
-
-    private ListenerConfiguration listenerConfiguration;
-
-    private SenderConfiguration senderConfiguration;
 
     private URI baseURI = URI.create(String.format("http://%s:%d", "localhost", 8490));
 
-    //@BeforeClass(groups = "passthroughPost",                        dependsOnGroups = "passthroughGET")
+    @BeforeClass
     public void setUp() {
-        listenerConfiguration = new ListenerConfiguration();
-        listenerConfiguration.setHost(TestUtil.TEST_HOST);
-        listenerConfiguration.setId("test-listener");
-        listenerConfiguration.setPort(TestUtil.TEST_ESB_PORT);
-        senderConfiguration = new SenderConfiguration("passthrough-sender");
-        httpTransportListener = TestUtil
-                .startCarbonTransport(listenerConfiguration, senderConfiguration, new PassthroughMessageProcessor());
-        httpServer = TestUtil.startHTTPServer(TestUtil.TEST_SERVER_PORT);
+        TransportsConfiguration configuration = YAMLTransportConfigurationBuilder
+                .build("src/test/resources/simple-test-config/netty-transports.yml");
+        httpTransportListener = TestUtil.startCarbonTransport(configuration, new PassthroughMessageProcessor());
+        httpServer = TestUtil.startHTTPServer(TestUtil.TEST_SERVER_PORT, testValue, Constants.TEXT_PLAIN);
     }
 
-    public PassThroughHttpPOSTMethodTestCase() {
-        super();
-    }
-
-    //@Test(groups = "passthroughPost",  dependsOnGroups = "passthroughGET")
-    public void passthroughPOSTTestCase() {
-        String testValue = "Test Message";
+    @Test
+    public void passthroughGetTest() {
         try {
-            HttpURLConnection urlConn = TestUtil.request(baseURI, "/", HttpMethod.POST.name(), true);
-            TestUtil.writeContent(urlConn, testValue);
+            HttpURLConnection urlConn = TestUtil.request(baseURI, "/", HttpMethod.GET.name(), true);
             String content = TestUtil.getContent(urlConn);
             assertEquals(testValue, content);
             urlConn.disconnect();
         } catch (IOException e) {
-            LOGGER.error("IO Exception occurred", e);
-            assertTrue(false);
+            log.error("Error while running passthroughGetTest ", e);
         }
-
     }
-    
-    //@AfterClass(groups = "passthroughPost", dependsOnGroups = "passthroughGET")
+
+    @Test
+    public void passthroughPostTest() {
+        try {
+            HttpURLConnection urlConn = TestUtil.request(baseURI, "/", HttpMethod.POST.name(), true);
+            String content = TestUtil.getContent(urlConn);
+            assertEquals(testValue, content);
+            urlConn.disconnect();
+        } catch (IOException e) {
+            log.error("Error while running passthroughPostTest ", e);
+        }
+    }
+
+    @AfterClass
     public void cleanUp() {
         TestUtil.cleanUp(httpTransportListener, httpServer);
     }
