+++ http/netty/components/org.wso2.carbon.transport.http.netty/src/test/resources/testng.xml
@@ -21,8 +21,8 @@
 <suite name="Transport test Suite">
     <test name="Transport test">
         <classes>
-            <class name="org.wso2.carbon.transport.http.netty.passthrough.PassThroughHttpGetMethodTestCase" />
-            <!--<class name="org.wso2.carbon.transport.http.netty.contentaware.ContentAwareMessageProcessorTestCase" />-->
+            <class name="org.wso2.carbon.transport.http.netty.passthrough.PassThroughHttpTestCase" />
+            <class name="org.wso2.carbon.transport.http.netty.contentaware.ContentAwareMessageProcessorTestCase" />
         </classes>
     </test>
 </suite>
